requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-bootstrap3==5.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-filter==0.10.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
fake-factory==0.5.1
gnureadline==6.3.3
gunicorn==19.3.0
hashID==3.1.4
hashids==1.1.0
ipython==3.1.0
matplotlib==1.4.3
nose==1.3.7
numpy==1.9.2
pandas==0.16.2
psycopg2==2.6.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
sqlparse==0.1.15
static3==0.6.1
Werkzeug==0.10.4
